Remove/rename more "helper" classes
