+++ tests/Calls/StaticCallsTest.php
@@ -7,8 +7,8 @@ use PHPStan\File\FileHelper;
 use PHPStan\Rules\Rule;
 use PHPStan\ShouldNotHappenException;
 use PHPStan\Testing\RuleTestCase;
+use Spaze\PHPStan\Rules\Disallowed\AllowedPath;
 use Spaze\PHPStan\Rules\Disallowed\DisallowedCallFactory;
-use Spaze\PHPStan\Rules\Disallowed\IsAllowedFileHelper;
 use Spaze\PHPStan\Rules\Disallowed\RuleErrors\DisallowedMethodRuleErrors;
 use Spaze\PHPStan\Rules\Disallowed\RuleErrors\DisallowedRuleErrors;
 use Spaze\PHPStan\Rules\Disallowed\Type\TypeResolver;
@@ -22,7 +22,7 @@ class StaticCallsTest extends RuleTestCase
 	protected function getRule(): Rule
 	{
 		return new StaticCalls(
-			new DisallowedMethodRuleErrors(new DisallowedRuleErrors(new IsAllowedFileHelper(new FileHelper(__DIR__))), new TypeResolver()),
+			new DisallowedMethodRuleErrors(new DisallowedRuleErrors(new AllowedPath(new FileHelper(__DIR__))), new TypeResolver()),
 			new DisallowedCallFactory(),
 			[
 				[
