+++ tests/Configs/InsecureConfigMethodCallsTest.php
@@ -8,9 +8,9 @@ use PHPStan\File\FileHelper;
 use PHPStan\Rules\Rule;
 use PHPStan\ShouldNotHappenException;
 use PHPStan\Testing\RuleTestCase;
+use Spaze\PHPStan\Rules\Disallowed\AllowedPath;
 use Spaze\PHPStan\Rules\Disallowed\Calls\MethodCalls;
 use Spaze\PHPStan\Rules\Disallowed\DisallowedCallFactory;
-use Spaze\PHPStan\Rules\Disallowed\IsAllowedFileHelper;
 use Spaze\PHPStan\Rules\Disallowed\RuleErrors\DisallowedMethodRuleErrors;
 use Spaze\PHPStan\Rules\Disallowed\RuleErrors\DisallowedRuleErrors;
 use Spaze\PHPStan\Rules\Disallowed\Type\TypeResolver;
@@ -26,7 +26,7 @@ class InsecureConfigMethodCallsTest extends RuleTestCase
 		// Load the configuration from this file
 		$config = Neon::decode(file_get_contents(__DIR__ . '/../../disallowed-insecure-calls.neon'));
 		return new MethodCalls(
-			new DisallowedMethodRuleErrors(new DisallowedRuleErrors(new IsAllowedFileHelper(new FileHelper(__DIR__))), new TypeResolver()),
+			new DisallowedMethodRuleErrors(new DisallowedRuleErrors(new AllowedPath(new FileHelper(__DIR__))), new TypeResolver()),
 			new DisallowedCallFactory(),
 			$config['parameters']['disallowedMethodCalls']
 		);
