+++ tests/Usages/NamespaceUsagesTypesTest.php
@@ -6,9 +6,9 @@ namespace Spaze\PHPStan\Rules\Disallowed\Usages;
 use PHPStan\File\FileHelper;
 use PHPStan\Rules\Rule;
 use PHPStan\Testing\RuleTestCase;
+use Spaze\PHPStan\Rules\Disallowed\AllowedPath;
 use Spaze\PHPStan\Rules\Disallowed\DisallowedNamespaceFactory;
-use Spaze\PHPStan\Rules\Disallowed\DisallowedNamespaceHelper;
-use Spaze\PHPStan\Rules\Disallowed\IsAllowedFileHelper;
+use Spaze\PHPStan\Rules\Disallowed\RuleErrors\DisallowedNamespaceRuleErrors;
 
 /**
  * @requires PHP > 8.1
@@ -19,7 +19,7 @@ class NamespaceUsagesTypesTest extends RuleTestCase
 	protected function getRule(): Rule
 	{
 		return new NamespaceUsages(
-			new DisallowedNamespaceHelper(new IsAllowedFileHelper(new FileHelper(__DIR__))),
+			new DisallowedNamespaceRuleErrors(new AllowedPath(new FileHelper(__DIR__))),
 			new DisallowedNamespaceFactory(),
 			[
 				[
